+++ tests/JsonWorks/Tests/Base.php
@@ -4,13 +4,9 @@ namespace JsonWorks\Tests;
 
 class Base extends \PHPUnit_Framework_TestCase
 {
-    public function validate($schema, $data)
+    protected function validate($schema, $data)
     {
-        $schema = json_decode(trim($schema));
-
-        if (null === $schema) {
-            throw new \InvalidArgumentException('Test not run, $schema not valid json');
-        }
+        $schema = $this->getSchemaObject($schema);
 
         if (is_string($data)) {
 
@@ -29,16 +25,9 @@ class Base extends \PHPUnit_Framework_TestCase
         return $validator->check($data, $schema);
     }
 
-    public function getDocument($schema, $data, $noException = false)
+    protected function getDocument($schema, $data, $noException = false)
     {
-
-        $schema = $schema ?: '{}';
-        $schema = json_decode(trim($schema));
-
-        if (null === $schema) {
-            throw new \InvalidArgumentException('Test not run, $schema not valid json');
-        }
-
+        $schema = $this->getSchemaObject($schema);
         $data = $data ?: null;
 
         if ($data) {
@@ -54,11 +43,67 @@ class Base extends \PHPUnit_Framework_TestCase
         return $document;
     }
 
-    public function callMethod($obj, $name, $args = array())
+    protected function callMethod($obj, $name, $args = array())
     {
         $class = new \ReflectionClass($obj);
         $method = $class->getMethod($name);
         $method->setAccessible(true);
         return $method->invokeArgs($obj, $args);
+    }
+
+    protected function getSchemaObject($schema)
+    {
+        $schema = $schema ?: '{}';
+
+        if (is_string($schema)) {
+            $schema = trim($schema);
+        }
+
+        return $this->fromJson($schema);
+    }
+
+    protected function fromJson($json)
+    {
+        $json = json_decode($json);
+
+        if (null === $json) {
+            throw new \InvalidArgumentException('Test not run, not valid json');
+        }
+
+        return $json;
+    }
+
+    protected function getExpectedJson($expected)
+    {
+        return json_encode(json_decode($expected));
+    }
+
+    protected function getFileExpectedJson($filename, $tabs = false)
+    {
+        if ($tabs) {
+            $data = file($filename, FILE_IGNORE_NEW_LINES);
+            return $this->fileSpacesToTabs($data);
+         } else {
+            return file_get_contents($filename);
+        }
+    }
+
+    protected function fileSpacesToTabs($data)
+    {
+        $space = str_repeat(chr(32), 4);
+
+        foreach($data as &$line)
+        {
+            $tabs = '';
+
+            while (0 === strpos($line, $space)) {
+                $line = substr($line, 4);
+                $tabs .= "\t";
+            }
+
+            $line = $tabs.$line;
+        }
+
+        return implode("\n", $data)."\n";
     }
 }